[ui] Add a "Pipelines" category containing pipeline templates in the node menu
Artificially add a "Pipelines" category in the node menu. The category
contains the list of available pipeline templates. Selecting a pipeline
instead of regular node type "creates" the pipeline in the GraphEditor
(using the "Import Project" functionality). The imported pipeline is
by default placed under the current graph.

This commit does not include enabling the search for pipelines from the
"Pipelines" category in the search bar.